# Replace JsonWebToken ReadPayloadValue with a delegate