--- a/src/Microsoft.IdentityModel.JsonWebTokens/Json/JsonWebToken.HeaderClaimSet.cs
+++ b/src/Microsoft.IdentityModel.JsonWebTokens/Json/JsonWebToken.HeaderClaimSet.cs
@@ -41,41 +41,7 @@ namespace Microsoft.IdentityModel.JsonWebTokens
             {
                 if (reader.TokenType == JsonTokenType.PropertyName)
                 {
-                    if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Alg))
-                    {
-                        _alg = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Alg, ClassName, true);
-                        claims[JwtHeaderParameterNames.Alg] = _alg;
-                    }
-                    else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Cty))
-                    {
-                        _cty = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Cty, ClassName, true);
-                        claims[JwtHeaderParameterNames.Cty] = _cty;
-                    }
-                    else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Kid))
-                    {
-                        _kid = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Kid, ClassName, true);
-                        claims[JwtHeaderParameterNames.Kid] = _kid;
-                    }
-                    else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Typ))
-                    {
-                        _typ = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Typ, ClassName, true);
-                        claims[JwtHeaderParameterNames.Typ] = _typ;
-                    }
-                    else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.X5t))
-                    {
-                        _x5t = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.X5t, ClassName, true);
-                        claims[JwtHeaderParameterNames.X5t] = _x5t;
-                    }
-                    else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Zip))
-                    {
-                        _zip = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Zip, ClassName, true);
-                        claims[JwtHeaderParameterNames.Zip] = _zip;
-                    }
-                    else
-                    {
-                        string propertyName = reader.GetString();
-                        claims[propertyName] = JsonSerializerPrimitives.ReadPropertyValueAsObject(ref reader, propertyName, JsonClaimSet.ClassName, true);
-                    }
+                    ReadHeaderValue(ref reader, claims);
                 }
                 // We read a JsonTokenType.StartObject above, exiting and positioning reader at next token.
                 else if (JsonSerializerPrimitives.IsReaderAtTokenType(ref reader, JsonTokenType.EndObject, false))
@@ -85,6 +51,57 @@ namespace Microsoft.IdentityModel.JsonWebTokens
             }
 
             return new JsonClaimSet(claims);
+        }
+
+
+        /// <summary>
+        /// Reads the value of a claim in the header and adds it to the <paramref name="claims"/> dictionary.
+        /// Can be overridden to read and add custom claims.
+        /// If a custom claim is read, the reader should be positioned at the next token after reading the claim.
+        /// </summary>
+        /// <param name="reader">The Utf8JsonReader instance positioned at a claim name token used to read the JSON payload.</param>
+        /// <param name="claims">Collection of claims that have been read from the reader.</param>
+        private protected virtual void ReadHeaderValue(ref Utf8JsonReader reader, IDictionary<string, object> claims)
+        {
+            if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Alg))
+            {
+                claims[JwtHeaderParameterNames.Alg] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Alg, ClassName, true);
+            }
+            else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Cty))
+            {
+                claims[JwtHeaderParameterNames.Cty] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Cty, ClassName, true);
+            }
+            else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Kid))
+            {
+                claims[JwtHeaderParameterNames.Kid] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Kid, ClassName, true);
+            }
+            else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Typ))
+            {
+                claims[JwtHeaderParameterNames.Typ] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Typ, ClassName, true); ;
+            }
+            else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.X5t))
+            {
+                claims[JwtHeaderParameterNames.X5t] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.X5t, ClassName, true);
+            }
+            else if (reader.ValueTextEquals(JwtHeaderUtf8Bytes.Zip))
+            {
+                claims[JwtHeaderParameterNames.Zip] = JsonSerializerPrimitives.ReadString(ref reader, JwtHeaderParameterNames.Zip, ClassName, true);
+            }
+            else
+            {
+                string propertyName = reader.GetString();
+
+                if (ReadTokenHeaderValueDelegates.TryGetValue(propertyName, out var readTokenHeaderValueDelegate) && readTokenHeaderValueDelegate != null)
+                {
+                    reader.Read();
+                    claims[propertyName] = readTokenHeaderValueDelegate(ref reader);
+                    reader.Read();
+                }
+                else
+                {
+                    claims[propertyName] = JsonSerializerPrimitives.ReadPropertyValueAsObject(ref reader, propertyName, JsonClaimSet.ClassName, true);
+                }
+            }
         }
     }
 }