--- a/src/Microsoft.IdentityModel.JsonWebTokens/Json/JsonWebToken.HeaderClaimSet.cs
+++ b/src/Microsoft.IdentityModel.JsonWebTokens/Json/JsonWebToken.HeaderClaimSet.cs
@@ -91,7 +91,7 @@ namespace Microsoft.IdentityModel.JsonWebTokens
             {
                 string propertyName = reader.GetString();
 
-                if (ReadTokenHeaderValueDelegates.TryGetValue(propertyName, out var readTokenHeaderValueDelegate) && readTokenHeaderValueDelegate != null)
+                if (ReadTokenPayloadValueDelegates.Count > 0 && ReadTokenHeaderValueDelegates.TryGetValue(propertyName, out var readTokenHeaderValueDelegate) && readTokenHeaderValueDelegate != null)
                 {
                     reader.Read();
                     claims[propertyName] = readTokenHeaderValueDelegate(ref reader);