--- a/src/Microsoft.IdentityModel.JsonWebTokens/JsonWebToken.cs
+++ b/src/Microsoft.IdentityModel.JsonWebTokens/JsonWebToken.cs
@@ -248,7 +248,7 @@ namespace Microsoft.IdentityModel.JsonWebTokens
         /// }
         /// </code>
         /// </remarks>
-        internal TryReadJwtClaim TryReadJwtClaim { get; set; }
+        private TryReadJwtClaim TryReadJwtClaim { get; set; }
 
         internal string ActualIssuer { get; set; }