… src/Microsoft.IdentityModel.JsonWebTokens/InternalAPI.Unshipped.txt
… src/Microsoft.IdentityModel.JsonWebTokens/JsonWebTokenHandler.cs
… src/Microsoft.IdentityModel.JsonWebTokens/PublicAPI.Unshipped.txt
… src/Microsoft.IdentityModel.JsonWebTokens/PublicAPI/net6.0/InternalAPI.Unshipped.txt
… src/Microsoft.IdentityModel.JsonWebTokens/PublicAPI/net8.0/InternalAPI.Unshipped.txt
… src/Microsoft.IdentityModel.JsonWebTokens/PublicAPI/net9.0/InternalAPI.Unshipped.txt
… src/Microsoft.IdentityModel.Tokens/Delegates.cs
… src/Microsoft.IdentityModel.Tokens/PublicAPI.Unshipped.txt
… src/Microsoft.IdentityModel.Tokens/TokenEnums.cs
… src/Microsoft.IdentityModel.Tokens/TokenValidationParameters.cs
… test/Microsoft.IdentityModel.JsonWebTokens.Tests/CustomJsonWebToken.cs
… test/Microsoft.IdentityModel.JsonWebTokens.Tests/JsonWebTokenTests.cs
… test/Microsoft.IdentityModel.TestUtils/ValidationDelegates.cs
… test/Microsoft.IdentityModel.Tokens.Tests/TokenValidationParametersTests.cs