Fixes #2982

Updated perf results: https://github.com/AzureAD/azure-activedirectory-identitymodel-extensions-for-dotnet/pull/2981#issuecomment-2749095467

This pull request introduces several updates and improvements to the `Microsoft.IdentityModel.JsonWebTokens` library, focusing on enhancing the handling of token payloads and custom claims. The most important changes include adding new constructors to the `JsonWebToken` class, updating the `JsonClaimSet` initialization, and introducing a new delegate for reading custom token payload values.

Enhancements to `JsonWebToken`:

* Added new constructors to the `JsonWebToken` class to support initializing tokens with custom delegates for reading token payload values. (`src/Microsoft.IdentityModel.JsonWebTokens/JsonWebToken.cs`) [[1]](diffhunk://#diff-a501634b44320ef31aa8a2eed25a471663f6ea2f3b5c9f6afe520c2c1203b7f3R86-R112) [[2]](diffhunk://#diff-a501634b44320ef31aa8a2eed25a471663f6ea2f3b5c9f6afe520c2c1203b7f3R137-R163) [[3]](diffhunk://#diff-a501634b44320ef31aa8a2eed25a471663f6ea2f3b5c9f6afe520c2c1203b7f3R195-R243)
* Introduced the `ReadTokenPayloadValueDelegates` property to the `JsonWebToken` class, allowing custom handling of specific claim names during token reading. (`src/Microsoft.IdentityModel.JsonWebTokens/JsonWebToken.cs`)

Updates to `TokenValidationParameters`:

* Added the `ReadTokenPayloadValueDelegates` property to the `TokenValidationParameters` class to support custom claim handling during token validation. (`src/Microsoft.IdentityModel.Tokens/TokenValidationParameters.cs`) [[1]](diffhunk://#diff-fd6fb46108ab0d6b00c019e12ed417bd02e026042ddaa215ec2b2301fc1304faR453-R458) [[2]](diffhunk://#diff-fd6fb46108ab0d6b00c019e12ed417bd02e026042ddaa215ec2b2301fc1304faL558-R565)
* Updated the copy constructor of `TokenValidationParameters` to include the new `ReadTokenPayloadValueDelegates` property. (`src/Microsoft.IdentityModel.Tokens/TokenValidationParameters.cs`)

Refactoring and cleanup:

* Refactored the `ReadPayloadValue` method in `JsonWebToken.PayloadClaimSet` to simplify the handling of standard claims and integrate custom delegate handling. (`src/Microsoft.IdentityModel.JsonWebTokens/Json/JsonWebToken.PayloadClaimSet.cs`) [[1]](diffhunk://#diff-517b23d070dc9135f820e9aac488d9af37a3b00b760550466b356a6438c1cbcaL42-R44) [[2]](diffhunk://#diff-517b23d070dc9135f820e9aac488d9af37a3b00b760550466b356a6438c1cbcaL81-R117)
* Removed the `CustomJsonWebToken` class from the tests, as its functionality is now covered by the new delegate-based approach. (`test/Microsoft.IdentityModel.JsonWebTokens.Tests/CustomJsonWebToken.cs`)

Delegate introduction:

* Introduced the `ReadTokenPayloadValueDelegate` delegate to handle custom claim reading during token payload processing. (`src/Microsoft.IdentityModel.Tokens/Delegates.cs`)