@pmaytak 2025-01-15
Updated results comparing delegates which have a dictionary as a parameter.

|Delegates| Method                                        | Mean     | Ratio | Error    | StdDev   | P90      | P95      | P100     | Ratio | Gen0   | Allocated | Alloc Ratio |
|-:|---------------------------------------------- |---------:|---------:|--:|---------:|---------:|---------:|---------:|------:|-------:|----------:|------------:|
|Without| JsonWebTokenHandler_ValidateTokenAsyncWithTVP | 30.13 μs | 1.00 | 0.169 μs | 0.374 μs | 30.61 μs | 30.74 μs | 31.07 μs |  1.00 | 0.2441 |   7.23 KB |        1.00 |
|With| JsonWebTokenHandler_ValidateTokenAsyncWithTVP | 30.22 μs | 1.00 | 0.315 μs | 0.679 μs | 31.14 μs | 31.48 μs | 33.30 μs |  1.00 | 0.3052 |   7.55 KB |        1.044 |

|Delegates| Method             | Mean     | Ratio | Error     | StdDev    | P90      | P95      | P100     | Gen0   | Allocated | Alloc Ratio |
|--:|------------------- |---------:|----------:|--: |----------:|---------:|---------:|---------:|-------:|----------:|--:|
|Without| ReadJWS_FromMemory | 7.523 μs | 1.00 | 0.0333 μs | 0.0724 μs | 7.599 μs | 7.612 μs | 7.672 μs | 0.2365 |   5.84 KB | 1.00 |
|With | ReadJWS_FromMemory | 7.370 μs | 0.980 |0.0182 μs | 0.0392 μs | 7.422 μs | 7.432 μs | 7.453 μs | 0.2441 |   6.06 KB | 1.038 |

BenchmarkDotNet v0.13.12, Windows 11 (10.0.26100.2605) (Hyper-V)
Intel Xeon Platinum 8370C CPU 2.80GHz, 1 CPU, 16 logical and 8 physical cores
.NET SDK 9.0.101
  [Host]    : .NET 8.0.11 (8.0.1124.51707), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI
  MediumRun : .NET 8.0.11 (8.0.1124.51707), X64 RyuJIT AVX-512F+CD+BW+DQ+VL+VBMI