@brentschmaltz 2025-02-18
It would be great to have the same design when reading the header.