@pmaytak 2025-03-24
The three scenarios compared below:
- Dev branch, JWT has custom claims, these are saved as string JSON.
- This feature branch, JWT has custom claims, no delegates are set, claims are saved as strings.
- This feature branch, JWT has custom claims and delegates which parse them into an object.
- Same cases as above, also creates claims.

| Branch  | Method                                                           | Mean     | Ratio | Error    | StdDev   | P90      | P95      | P100     | Gen0   | Allocated | Alloc Ratio |
|---------|------------------------------------------------------------------|----------|-------|----------|----------|----------|----------|----------|--------|-----------|-------------|
| Base    | JsonWebTokenHandler_ValidateTokenAsync                           | 30.59 μs | **1.000**  | 0.037 μs | 0.079 μs | 30.68 μs | 30.72 μs | 30.85 μs | 0.4272 | 7.51 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync                           | 31.27 μs | **1.022** | 0.322 μs | 0.672 μs | 32.30 μs | 32.40 μs | 32.61 μs | 0.4272 | 7.94 KB   | **1.057**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates              | 31.47 μs | **1.029**  | 0.110 μs | 0.242 μs | 31.79 μs | 31.87 μs | 32.03 μs | 0.4883 | 8 KB      | **1.065**        |
| Base    | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 35.38 μs | **1.000**  | 0.130 μs | 0.286 μs | 35.80 μs | 35.88 μs | 36.10 μs | 1.0376 | 17.3 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 35.20 μs | **0.995** | 0.057 μs | 0.121 μs | 35.33 μs | 35.36 μs | 35.43 μs | 1.0376 | 17.73 KB  | **1.025**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates_CreateClaims | 35.64 μs | **1.007**  | 0.069 μs | 0.151 μs | 35.84 μs | 35.86 μs | 36.00 μs | 1.0376 | 17.79 KB  | **1.028**       |