@pmaytak 2025-03-24
The three scenarios compared below:
- Dev branch, JWT has custom claims, these are saved as string JSON.
- This feature branch, JWT has custom claims, no delegates are set, claims are saved as strings.
- This feature branch, JWT has custom claims and delegates which parse them into an object.
- Same cases as above, also creates claims.

| Branch  | Method                                                           | Mean     | Ratio | Error    | StdDev   | P90      | P95      | P100     | Gen0   | Allocated | Alloc Ratio |
|---------|------------------------------------------------------------------|----------|-------|----------|----------|----------|----------|----------|--------|-----------|-------------|
| Base    | JsonWebTokenHandler_ValidateTokenAsync                           | 30.59 μs | **1.000**  | 0.037 μs | 0.079 μs | 30.68 μs | 30.72 μs | 30.85 μs | 0.4272 | 7.51 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync                           | 30.29 μs | **0.990** | 0.061 μs | 0.131 μs | 30.47 μs | 30.51 μs | 30.65 μs | 0.4272 | 7.63 KB   | **1.016**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates              | 31.16 μs | **1.019** | 0.099 μs | 0.217 μs | 31.51 μs | 31.54 μs | 31.59 μs | 0.4272 | 7.77 KB   | **1.035**        |
| Base    | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 35.38 μs | **1.000**  | 0.130 μs | 0.286 μs | 35.80 μs | 35.88 μs | 36.10 μs | 1.0376 | 17.3 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 34.65 μs | **0.979**  | 0.074 μs | 0.162 μs | 34.84 μs | 34.93 μs | 35.03 μs | 0.9766 | 17.41 KB  | **1.006**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates_CreateClaims | 35.70 μs | **1.009**  | 0.083 μs | 0.180 μs | 35.86 μs | 35.96 μs | 36.47 μs | 1.0376 | 17.55 KB  | **1.008**        |