@pmaytak 2025-04-09
The difference from the previous commit is that there's now one delegate instead of a dictionary.

The three scenarios compared below:
- Dev branch, JWT has custom claims, these are saved as string JSON.
- This feature branch, JWT has custom claims, no delegates are set, claims are saved as strings.
- This feature branch, JWT has custom claims and delegates which parse them into an object.
- Same cases as above, also creates claims.

| Branch  | Method                                                           | Mean     | Ratio | Error    | StdDev   | P90      | P95      | P100     | Gen0   | Allocated | Alloc Ratio |
|---------|------------------------------------------------------------------|----------|-------|----------|----------|----------|----------|----------|--------|-----------|-------------|
| Base    | JsonWebTokenHandler_ValidateTokenAsync                           | 30.59 μs | **1.000**  | 0.037 μs | 0.079 μs | 30.68 μs | 30.72 μs | 30.85 μs | 0.4272 | 7.51 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync                           | 30.56 μs |  **1.000**  | 0.055 μs | 0.120 μs | 30.71 μs | 30.73 μs | 30.87 μs | 0.4272 |   7.61 KB |  **1.013** |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates              | 31.08 μs | **1.016**  | 0.075 μs | 0.166 μs | 31.30 μs | 31.33 μs | 31.53 μs  | 0.4272 |   7.75 KB | **1.032**  |
| Base    | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 35.38 μs | **1.000**  | 0.130 μs | 0.286 μs | 35.80 μs | 35.88 μs | 36.10 μs | 1.0376 | 17.3 KB   | **1.000**        |
| Feature | JsonWebTokenHandler_ValidateTokenAsync_CreateClaims              | 35.34 μs | **0.999**  | 0.087 μs | 0.189 μs | 35.54 μs | 35.61 μs | 35.81 μs | 0.9766 |   17.4 KB | **1.006**  |
| Feature | JsonWebTokenHandler_ValidateTokenAsyncWithDelegates_CreateClaims | 35.25 μs | **0.996**  | 0.072 μs | 0.159 μs | 35.47 μs | 35.49 μs | 35.59 μs | 1.0376 |  17.54 KB | **1.014**  |